Pause before each websocket send, and some requests
This slows down playback to match the "think times" recorded in
the original run. Each websocket-send pauses the same amount of
time that elapsed between its start, and the previous event's
start, in the original run. This may be an oversimplified
heuristic for determining the amount of time to wait. For one
thing, we should probably measure from the previous event's *end*
time, not start time. But we don't record that today.

Regular HTTP requests do not pause, unless they occur after the
websocket connection has been established. I'm assuming that the
requests before WS-open are all related to initial page load, and
the ones afterwards are related to user actions (like clicking a
downloadButton, or paging through a DT table).

Also added a comment to the top of recordings to indicate the
target URL. This is helpful in case you don't bother to use good
filenames for your recording files.

Also, now recording HTTP request response codes, so benign
failures (like 404 on favicon.ico) don't cause the playback to
fail.