Fix SSP PAM-auth'd app playback
While cookies retained from the first request to the app URL were retained,
they weren't actually being sent by Apache HttpClient on the form POST
because their Domains weren't set. This resulted in the message "invalid csrf
token" and a 403 response from /__login__

The fix was to ensure cookies have a domain that matches the request they will
be included in.

I don't know why this only affected SSP with PAM and not SSP with file auth.